+++ src/python/pants/backend/tools/fmt/preamble/register.py
@@ -1,7 +1,7 @@
 # Copyright 2022 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.tools.fmt.preamble import rules as preamble_rules
+from pants.backend.tools.fmt.preamble import rules as preamble_rules
 
 
 def rules():
